pyproject.toml
[project]
name = "image_processing"
version = "0.1.0"
description = "A Python library for basic image processing tasks."
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow~10.0.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 88

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
